This is a document file for the texas TI-nspire CX CAS
This file needs to be downloaded to your local computer then opened into the TI-Nspire CX CAS Student Software in order to be saved afterwards to your calculator, the formulas will be available once you open the file in your calculator. Once in the calculator then you can click into catalogue (small book icon)--> click on 6 to open the libraries --> select current problem, and this will display the available formulas.